file move preview for 419
Hey, this is how the files will look for 419. There will be some other new files in the 'search' module too, refactoring the predicate panels over. This is a preview and also guidance for github so it can better track the history of the moved files once I commit all changes.