Speed up build process by caching artifacts
because the pyoxydizer isn't updated that often, why build it every time and waste 20 min. This change caches the built artifacts, so the pyoxydizer step runs in matter of seconds. Also removed release job, since it has nothing to do in here. Every component has their build job, but release should be it's own dedicated job, since it relies on many parts.