Fix the application icon on KDE Plasma Wayland
This fixes the application icon on KDE Plasma Wayland (tested on Plasma
6) and possibly other Wayland compositors too.

The desktop filename was chosen because this is the identifier that the
Hydrus Flatpak application uses.

Distribution packagers should make sure that this name matches with the
desktop file that they install (for example, Gentoo installs a
/usr/share/applications/hydrus-client-hydrus.desktop file).

It is strongly suggested that distributions install their Hydrus
.desktop file at
/usr/share/applications/io.github.hydrusnetwork.hydrus.desktop to avoid
needing to apply extra patches, etc.