Update bug_report.yaml (#1635)
We have not had a reported issue from macOS 11 or earlier in over a year. We have had recent reports from someone choosing the macOS 12 option, but in one of their issues, they admit to using a Macbook with an M3 chip, which released after macOS 13 was released. So, they probably just chose the most recent macOS version from the dropdown, rather than choosing "macOS other".