Explain the random sample behavior in complex system:limit searches (#1760)
I found myself a bit confused by the wording in the tooltip in the client, and didn't find that the manual clarified anything. I don't have any good ideas for how the tooltip text should be changed to be clearer, but at least this adds a concrete explanation to the manual.