Non-Recursive Folder Search for local import (#1780)
* Update ClientFiles.py

Added Check for non-recursive folder searches that causes folders inside of non-recursive folder imports to be ignored.

* Update ClientGUIScrolledPanelsReview.py

Modified AddFolder function to turn the path added to queue into a dictionary that contains the path to be added to queue and a key that lets the GetAllPaths function know not to go into any subfolders.

* Converted call to lambda for non-recursive import

* Fixed incorrect path definition